Remove review marks step from interview portal
